pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["gomoku_rl"]

[project]
name = "gomoku_rl"
version = "0.0.0"
authors = [{ name = "Sicheng He", email = "hesicheng20@gmail.com" }]
description = "train AI agents to play Free-style Gomoku."
readme = "README.md"
keywords = ["RL", "pytorch"]
dependencies = [
    "torch",
    "torchrl==0.2.1",
    "scipy",
    "omegaconf",
    "hydra-core",
    "tqdm",
    "wandb",
    "matplotlib",
]
requires-python = ">=3.10"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]

[project.optional-dependencies]
gui = ["PyQt5"]
test = ["pytest"]
